Испоавид проблему с уменьшением числа товаров возле значка корзины в хедере, связанную с возможностью уменьшать кол-во конкретного товара, в момент, когда происходит удаление этого товара из бд(конда товара становится меньше 1)

+++ src/components/v-cart-item.vue
@@ -85,9 +85,12 @@ export default {
         increment(-1);
         emit("productPrice", -product.todayprice);
       } else {
-        increment(-1);
-        emit("productPrice", -product.todayprice);
-        deleteCartItem(product);
+        if (quantityProd.value - 1 >= 0) {
+          quantityProd.value -= 1;
+          increment(-1);
+          emit("productPrice", -product.todayprice);
+          deleteCartItem(product);
+        }
       }
     };
